Modified for running CG-QCT with 4 Atoms system

+++ scripts/executing/Clean.sh
@@ -26,51 +26,56 @@ function RmAll {
   export TranFlg=${1}
   export Tran=${2}
   export Tint=${3}
-  export MinLevel1=${4}
-  export MaxLevel1=${5}
-  
-  echo "  [Clean]: COARSEAIR_OUTPUT_DIR  = "${COARSEAIR_OUTPUT_DIR}
-  echo "  [Clean]: TranFlg               = "${TranFlg}
-  echo "  [Clean]: Tran                  = "${Tran}
-  echo "  [Clean]: Tint                  = "${Tint}
-  echo "  [Clean]: MinLevel1             = "${MinLevel1}
-  echo "  [Clean]: MaxLevel1             = "${MaxLevel1}
-  echo "  [Clean]: MinLevel2             = "${MinLevel2}
-  echo "  [Clean]: MaxLevel2             = "${MaxLevel2}
+  export NMolecules=${4}
+  export MinLevel1=${5}
+  export MaxLevel1=${6}
+  export MinLevel1=${7}
+  export MaxLevel1=${8}
+
+  echo "  [RmAll]: COARSEAIR_OUTPUT_DIR  = "${COARSEAIR_OUTPUT_DIR}
+  echo "  [RmAll]: TranFlg               = "${TranFlg}
+  echo "  [RmAll]: Tran                  = "${Tran}
+  echo "  [RmAll]: Tint                  = "${Tint}
+  echo "  [RmAll]: NMolecules            = "${NMolecules}
+  echo "  [RmAll]: MinLevel1             = "${MinLevel1}
+  echo "  [RmAll]: MaxLevel1             = "${MaxLevel1}
+  echo "  [RmAll]: MinLevel2             = "${MinLevel2}
+  echo "  [RmAll]: MaxLevel2             = "${MaxLevel2}
    
 
-  iLevels1=${MinLevel1}
-  while [ ${iLevels1} -le ${MaxLevel1} ]
-    do
-    echo "  [Clean]: --- Molecule 1, Level/Bin " ${iLevels1} " ----------------------------- "
-
-    iLevels2=0  
-    #iLevels2=${MinLevel2}
-    #while [ ${iLevels2} -le ${MaxLevel2} ]; do
-    #  echo "  [Clean]: ----- Molecule 2, Level/Bin = " ${iLevels2} " --------------------- "
-    #  echo "  [Clean]"
-      
-      
-      if [ ${TranFlg} -eq 0 ]; then 
-        export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"E_"${Tran%.*}"_T_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
-      else
-        export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"T_"${Tran%.*}"_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
+  MinLevel1Temp=${MinLevel1}
+  MaxLevel1Temp=${MaxLevel1}
+  MinLevel2Temp=${MinLevel2}
+  MaxLevel2Temp=${MaxLevel2}
+  for (( iLevel1=1; iLevel1<=${NLevels1}; iLevel1++ )); do
+    if [ ${iLevel1} -ge ${MinLevel1Temp} ] && [ ${iLevel1} -le ${MaxLevel1Temp} ]; then
+      if [ ${MinLevel2} -eq 0 ] && [ ${NMolecules} -gt 1 ]; then
+        MinLevel2Temp=${iLevel1}
       fi
+      for (( iLevel2=1; iLevel2<=${NLevels2}; iLevel2++ )); do
+        if [ ${iLevel2} -ge ${MinLevel2Temp} ] && [ ${iLevel2} -le ${MaxLevel2Temp} ]; then
+          echo "  [RmAll]: --- Molecule 1, Level/Bin " ${iLevels1} " ----------------------------- "
+          echo "  [RmAll]: ----- Molecule 2, Level/Bin = " ${iLevels2} " --------------------- "
+          echo "  [RmAll]"
+
       
-      
-      #if [ -f ${COARSEAIR_BIN_OUTPUT_DIR} ]; then
-      #  echo ${COARSEAIR_BIN_OUTPUT_DIR}
-        rm -rf ${COARSEAIR_BIN_OUTPUT_DIR}
-      #fi
-    
-    
-    #  echo "  [Clean]: ---------------------------------------------------------- "
-    #  iLevels2=$((iLevels2+1))
-    #done
-    
-    echo "  [Clean]: ------------------------------------------------------------ "
+          if [ ${TranFlg} -eq 0 ]; then 
+            export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"E_"${Tran%.*}"_T_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
+          else
+            export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"T_"${Tran%.*}"_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
+          fi
+
+
+          #if [ -f ${COARSEAIR_BIN_OUTPUT_DIR} ]; then
+          #  echo ${COARSEAIR_BIN_OUTPUT_DIR}
+          rm -rf ${COARSEAIR_BIN_OUTPUT_DIR}
+          #fi
+        fi
+        echo "  [RmAll]: ---------------------------------------------------------- "
+      done
+    fi
+    echo "  [RmAll]: ------------------------------------------------------------ "
     echo " "
-    iLevels1=$((iLevels1+1))
   done         
       
 }
@@ -85,52 +90,59 @@ function Clean {
   export TranFlg=${1}
   export Tran=${2}
   export Tint=${3}
-  export MinLevel1=${4}
-  export MaxLevel1=${5}
+  export NMolecules=${4}
+  export MinLevel1=${5}
+  export MaxLevel1=${6}
+  export MinLevel1=${7}
+  export MaxLevel1=${8}
   
   echo "  [Clean]: COARSEAIR_OUTPUT_DIR  = "${COARSEAIR_OUTPUT_DIR}
   echo "  [Clean]: TranFlg               = "${TranFlg}
   echo "  [Clean]: Tran                  = "${Tran}
   echo "  [Clean]: Tint                  = "${Tint}
+  echo "  [Clean]: NMolecules            = "${NMolecules}
   echo "  [Clean]: MinLevel1             = "${MinLevel1}
   echo "  [Clean]: MaxLevel1             = "${MaxLevel1}
   echo "  [Clean]: MinLevel2             = "${MinLevel2}
   echo "  [Clean]: MaxLevel2             = "${MaxLevel2}
 
 
-  iLevels1=${MinLevel1}
-  while [ ${iLevels1} -le ${MaxLevel1} ]
-    do
-    echo "  [Clean]: --- Molecule 1, Level/Bin " ${iLevels1} " ----------------------------- "
+  MinLevel1Temp=${MinLevel1}
+  MaxLevel1Temp=${MaxLevel1}
+  MinLevel2Temp=${MinLevel2}
+  MaxLevel2Temp=${MaxLevel2}
+  for (( iLevel1=1; iLevel1<=${NLevels1}; iLevel1++ )); do
+    if [ ${iLevel1} -ge ${MinLevel1Temp} ] && [ ${iLevel1} -le ${MaxLevel1Temp} ]; then
+      if [ ${MinLevel2} -eq 0 ] && [ ${NMolecules} -gt 1 ]; then
+        MinLevel2Temp=${iLevel1}
+      fi
+      for (( iLevel2=1; iLevel2<=${NLevels2}; iLevel2++ )); do
+        if [ ${iLevel2} -ge ${MinLevel2Temp} ] && [ ${iLevel2} -le ${MaxLevel2Temp} ]; then
+          echo "  [Clean]: --- Molecule 1, Level/Bin " ${iLevels1} " ----------------------------- "
+          echo "  [Clean]: ----- Molecule 2, Level/Bin = " ${iLevels2} " --------------------- "
+          echo "  [Clean]"
 
-    iLevels2=0  
-    #iLevels2=${MinLevel2}
-    #while [ ${iLevels2} -le ${MaxLevel2} ]; do
-    #  echo "  [Clean]: ----- Molecule 2, Level/Bin = " ${iLevels2} " --------------------- "
-    #  echo "  [Clean]"
       
-      
-      if [ ${TranFlg} -eq 0 ]; then 
-        export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"E_"${Tran%.*}"_T_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
-      else
-        export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"T_"${Tran%.*}"_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
-      fi
-      
-      
-      if [ -f ${COARSEAIR_BIN_OUTPUT_DIR}/NConvTraj.dat ]; then
-        rm -rf ${COARSEAIR_BIN_OUTPUT_DIR}/Node*
-        rm -rf ${COARSEAIR_BIN_OUTPUT_DIR}/statistics-*
-      fi
-    
-    
-    #  echo "  [Clean]: ---------------------------------------------------------- "
-    #  iLevels2=$((iLevels2+1))
-    #done
-    
+          if [ ${TranFlg} -eq 0 ]; then 
+            export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"E_"${Tran%.*}"_T_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
+          else
+            export COARSEAIR_BIN_OUTPUT_DIR=${COARSEAIR_OUTPUT_DIR}/"T_"${Tran%.*}"_"${Tint%.*}/"Bins_"${iLevels1}"_"${iLevels2}
+          fi
+
+
+          if [ -f ${COARSEAIR_BIN_OUTPUT_DIR}/NConvTraj.dat ]; then
+            rm -rf ${COARSEAIR_BIN_OUTPUT_DIR}/Node*
+            rm -rf ${COARSEAIR_BIN_OUTPUT_DIR}/statistics-*
+          fi
+
+
+        fi
+        echo "  [Clean]: ---------------------------------------------------------- "
+      done
+    fi
     echo "  [Clean]: ------------------------------------------------------------ "
     echo " "
-    iLevels1=$((iLevels1+1))
-  done         
+  done     
       
 }
 #================================================================================================================================#
